\name{mirtarbase}
\alias{mirtarbase}
\title{
  miRTarBase (MTI) database package.
}
\description{
  This package contains the miRNA-target gene interactions (MTI)
  provided by the miRTarBase.
}
\usage{
  mirtarbase( prefix="" )
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{prefix}{
    Character string that could be pre-pended to the individual
    informations printed.
  }
}
\value{
  Prints some information for the package.
}
\author{
Johannes Rainer
}
\keyword{ data }
